Migrate attendance code to TS
